EnumValue as interface #1053
