+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/models/EnumValueImpl.java
@@ -1,0 +1,23 @@
+package io.sphere.sdk.models;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+final class EnumValueImpl extends Base implements EnumValue {
+    private final String key;
+    private final String label;
+
+    @JsonCreator
+    EnumValueImpl(final String key, final String label) {
+        this.key = key;
+        this.label = label;
+    }
+
+    @Override
+    public String getKey() {
+        return key;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+}
